+++ pkg/nn/init.go
@@ -45,9 +45,9 @@ func getArchitecture(name string) NeuralNetwork {
 }
 
 // getRand return random number from -0.5 to 0.5
-func getRandFloat() (r float32) {
+func getRandFloat() (r float64) {
 	for r == 0 || r > .5 || r < -.5 {
-		r = float32(rand.Float64() - .5)
+		r = rand.Float64() - .5
 	}
 	return
 }
